# app/Http/Controllers/StatusLeitura/PesquisarStatusLeituraController.php
<?php

namespace App\Http\Controllers\StatusLeitura;

use App\Http\Controllers\Controller;
use App\Http\Resources\StatusLeitura\StatusLeituraResource;
use App\Http\Services\StatusLeitura\StatusLeiturasPesquisa;

class PesquisarStatusLeituraController extends Controller
{
    public function __construct(
        protected StatusLeiturasPesquisa $serviceStatusLeiturasPesquisa
    ) {}

    /**
     * @OA\Get(
     *   path="/api/status/leituras",
     *   summary="Pesquisar Status da Leitura",
     *   description="Pesquisar Status da Leitura",
     *   tags={"Status Leituras"},
     *   security={{"bearerAuth":{}}},
     *
     *   @OA\Parameter(
     *      name="token",
     *                                         in="query",
     *      required=true,
     *      description="Token de autenticação do usuário",
     *
     *      @OA\Schema(type="string")
     *   ),
     *
     *   @OA\Response(
     *     response=200,
     *     description="Pesquisar Status da Leitura",
     *
     *     @OA\JsonContent(
     *       type="object",
     *
     *       @OA\Property(property="success", type="boolean"),
     *       @OA\Property(property="message", type="string"),
     *       @OA\Property(property="data", type="object", description="Dados adicionais, se houver")
     *     )
     *   ),
     *
     *   @OA\Response(
     *     response=401,
     *     description="Não autorizado, token inválido ou ausente"
     *   ),
     *   @OA\Response(
     *     response=500,
     *     description="Erro interno no servidor"
     *   )
     * )
     */
    public function __invoke()
    {
        return StatusLeituraResource::collection(
            $this->serviceStatusLeiturasPesquisa->listarStatusLeitura()
        );
    }
}