# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Minhas Leituras Documentação",
        "description": "Gerenciamento de Leituras Este projeto é uma aplicação de gerenciamento de leituras onde os usuários podem cadastrar, organizar e acompanhar seu progresso de leitura de livros, HQs e outros materiais. O sistema permite que os usuários gerenciem suas leituras, definam metas, escrevam avaliações, recebam lembretes para ler, e muito mais.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/avaliacoes/{id_avaliacao_leitura}": {
            "delete": {
                "tags": [
                    "Avaliacao Leituras"
                ],
                "summary": "Excluir Avaliação Leitura do Usuário",
                "description": "Excluir Avaliação Leitura do Usuário",
                "operationId": "479f4620eedeb85585159eda06f8ec87",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token de autenticação do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id_avaliacao_leitura",
                        "in": "path",
                        "description": "Id Avaliacao Leitura da Avaliação",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Leitura excluída com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "Dados adicionais, se houver",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado, token inválido ou ausente"
                    },
                    "404": {
                        "description": "Leitura não encontrada"
                    },
                    "500": {
                        "description": "Erro interno no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/leituras/isbn-api/{isbn}": {
            "get": {
                "tags": [
                    "Leituras"
                ],
                "summary": "Pesquisar Leitura pelo ISBN em API externa",
                "description": "Pesquisar Leituras pelo ISBN em API externa",
                "operationId": "d46f2fa08188a5eadd37eb2e27ecaca0",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token de autenticação do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isbn",
                        "in": "path",
                        "description": "ISBN do livro a ser pesquisado",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pesquisar Leituras com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "Dados adicionais, se houver",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado, token inválido ou ausente"
                    },
                    "500": {
                        "description": "Erro interno no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/leituras/aleatoria": {
            "get": {
                "tags": [
                    "Leituras"
                ],
                "summary": "Pesquisar Leitura Aleatória",
                "description": "Pesquisar Leitura Aleatória",
                "operationId": "c78e33d606bea26a347953f141a16c55",
                "responses": {
                    "200": {
                        "description": "Pesquisar Leitura Aleatória",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "Dados adicionais, se houver",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado, token inválido ou ausente"
                    },
                    "500": {
                        "description": "Erro interno no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/progresso/total": {
            "post": {
                "tags": [
                    "Progresso Leituras"
                ],
                "summary": "Pesquisa Leitura Progresso Total",
                "description": "Pesquisa Leitura Progresso Total",
                "operationId": "c9317dd7cdc491d27ee0403992b920dd",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token de autenticação do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id_leitura"
                                ],
                                "properties": {
                                    "id_leitura": {
                                        "description": "ID da leitura",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pesquisa Leitura Progresso Total realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "Dados adicionais, se houver",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado, token inválido ou ausente"
                    },
                    "500": {
                        "description": "Erro interno no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/status/leituras": {
            "get": {
                "tags": [
                    "Status Leituras"
                ],
                "summary": "Pesquisar Status da Leitura",
                "description": "Pesquisar Status da Leitura",
                "operationId": "f4988959276fe29e77984b9fc3a52aac",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token de autenticação do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pesquisar Status da Leitura",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "Dados adicionais, se houver",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado, token inválido ou ausente"
                    },
                    "500": {
                        "description": "Erro interno no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}